# make CI tests faster (2.4X faster on local OSX), and make gcleak2 leak detection more accurate