## before PR

there's a single test that takes a whopping 335 seconds (on my local OSX; on other machines may be different), representing > 57% of CI tests time (585 s) (ie, all of `testament/testament all`; this excludes other things in `./koch runCI`).

on local OSX:
```
XDG_CONFIG_HOME= nim c -r testament/testament --nim:./bin/nim all
PASS: tests/gc/gcleak2 C --gc:boehm                                (165.00503802 secs)
PASS: tests/gc/gcleak2 C -d:release --gc:boehm                     (170.92052507 secs)
...
```

## after PR
* these:
./bin/nim c -r -d:release --gc:boehm tests/gc/gcleak2.nim
./bin/nim c -r --gc:boehm tests/gc/gcleak2.nim
now both run in 1 sec, by simply reducing number of iterations in this case; the test is in fact more useful than before despite smaller numIter, as we check that the occupied memory doesn't vary after few iterations; so running more iterations would make no difference.

* other deterministic gc tested (eg: --gc:orc, --gc:arc) also are tested like --gc:boehm (ie fixed point at a few iterations)

* for the other gc tested (ie `./bin/nim c -r tests/gc/gcleak2.nim` + other gc) , we also modify the test to make it more useful, by checking that occupied memory follows a set of cycles, each with identical min/max values after a few cycles. The test before this PR was more rough, and would not have detected a small leak accumulating since it just had a coarse upperbound on memory use.

and overall, `testament/testament all` now runs in 242s, a 2.4X speedup (on my local OSX)

## Note
on azure CI machines, the overall speedup is less significant compared to the speedup observed on my local OSX machine (OSX catalina 10.15, 16GB RAM 2.3 GHz 8-Core Intel Core i9); for some reason `./bin/nim c -r --gc:boehm tests/gc/gcleak2.nim` was 165s (before this PR) on my local machine but only 65s on azure OSX machines (note: azure uses 10.14 not 10.15, could be a factor)  eg:
https://dev.azure.com/nim-lang/255dfe86-e590-40bb-a8a2-3c0295ebdeb1/_apis/build/builds/2835/logs/60
```
2020-02-24T22:22:05.2156800Z [32mPASS: [36mtests/gc/gcleak2 C --gc:boehm                               [34m (65.62197399 secs)[0m
2020-02-24T22:22:05.2157710Z [32mPASS: [36mtests/gc/gcleak2 C -d:release --gc:boehm                    [34m (54.69007993 secs)[0m
```
still, this PR saves 120s on each such azure machine, and it saves 335s (a 2.4X overall speedup) on a local OSX machine, and makes the tests more accurate.