@narimiran 2020-02-25
> (2.4X faster on local OSX)

> on azure CI machines, the overall speedup is less significant

> was (...) only 65s on azure OSX machines 

> still, this PR saves 120s on each such azure machine

Ok, I'll be @krux02 this time: can you give us a straight-forward non-clickbait realistic numbers for our CIs? (We don't care about your local machine numbers)