@timotheecour 2020-02-25
in absolute terms, on azure CI machines, it saves 120s per machine; it's the slowest test

> We don't care about your local machine numbers

unless other people running latest OSX (10.15) have same issue of  `nim c -r -d:release --gc:boehm tests/gc/gcleak2.nim` being so slow; I'd like someone with 10.15 to report the runtime for that command so I know it's not just my machine