@Araq 2020-02-26
But now the test is much harder to understand. Why not simply change the number 1_000_000 to 10_000 instead and call it day? It used to be green with one million for months if not for years, no need to spend so much effort on it.