@krux02 2020-02-26
First of all, thank you for tackling this problem. This is really an important problem to solve. Slow CI is something that affects everybody and faster CI feedback is oil in the Nim development machinery. So for me a PR tacking this problem is high priority.

But what I don't like is, might be subjective but I like minimalism when it comes to the introduction of `if` and `when` branches, and you've introduced three `when isDeterministic` alone. I also really don't like variable declarations behind a `when` statement, as they create a context where variables might not be declared and you have to use more `when` statements later on (as you did in this pr). I think you can do better about it. And here I think your can use compile time polymorphism effeciently. 

An idea that would be this:

```nim
var tester: (when isDeterministic: DeterministicTester else: HeuristicTester)

type 
  HeuristicTester = object
    numCycle
    memPrevious: int
    memMin: int
    numCollections: int
  [...]

proc measure(tester: HeuristicTester; mem: int) = [...]
proc measure(tester: DeterministicTester; mem: int) = [...]

[...]
   tester.measure(mem) # instead of when isDeterministinc 

```