@Clyybber 2020-02-26
I don't think this extra complexity is worth it, just to save 2 minutes.