@narimiran 2020-02-26
> but 2 minutes seems as a good win?

...which might be achieved by changing one number in one line, as shown above, without the extra complexity.