@timotheecour 2020-02-27
PTAL
* reverted changes (maybe don't squash so at least we have it in git history)
* then re-enabled simpler change that gives same speedup (335s on local OSX, 120 on azure machine)
* then did similar modification for ./tests/gc/gcleak.nim (saves another  33s on local OSX)

## note
yes the original changes were more complex, but they were also more precise, checking for loop/cycle invariants as opposed to current situation where we're just checking for a very coarse 1-size-fits-all upperbound on used memory (across OS/CPU/optimization level; and especially across GC algorithms despite them being widely different)
Whereas the test I had written originally checked that:
* for some GC algos, memory usage was constant per iteration i after i >= 3
* for the other algos, memory usage followed a set of cycles with constant min/max memory usage after seeing 5 cycles

Which should find even the smallest leak should that occur in that code.

> It used to be green with one million for months if not for years, no need to spend so much effort on it.

the problem isnt' that it's green; the problem is that it's too loose to detect some leaks. Leaks do happen eg:
* https://github.com/nim-lang/Nim/issues/13080
* https://github.com/nim-lang/Nim/pull/10498 etc